Rename dg_poly_spaces.py to poly_spaces.py.
