+++ sfepy/discrete/dg/gen_legendre_simplex_base.py
@@ -15,7 +15,7 @@ from operator import mul
 import numpy as np
 
 from sfepy.base.ioutils import InDir
-from sfepy.discrete.dg.dg_poly_spaces import iter_by_order
+from sfepy.discrete.dg.poly_spaces import iter_by_order
 
 helps = {
     'max_order' :
